+++ src/types.rs
@@ -38,6 +38,7 @@ mod private {
     impl Sealed for tables::Customer {}
     impl Sealed for tables::DataSource {}
     impl Sealed for tables::Filter {}
+    impl Sealed for tables::Host {}
     impl Sealed for tables::InnerNode {}
     impl Sealed for tables::Model {}
     impl Sealed for tables::ModelIndicator {}